refactor(models): widen integer response fields int -> long
ESI's OpenAPI schema types every integer in a 200/201 response body as
int64 (verified: 661/661 integer leaves, zero int32), and EVE's own IDs
already sit on the int32 boundary. SpecCheck flagged 496 of these.

- Every `int` property on a response model is now `long`; `int[]` -> `long[]`.
- AuthorizedCharacterData.CharacterID / AllianceID / CorporationID /
  FactionID widened too (SsoLogic parses the subject claim as long now).
- Methods that return a bare id list/scalar now return EsiResponse<long[]>
  / EsiResponse<long>: Alliance.All/Corporations, Clones.Implants,
  Contacts.Add, Corporation.NpcCorps/Members/MemberLimit,
  Dogma.Attributes/Effects, Mail.New, Market.Groups/Types, Routes.Map,
  Wars.All, and the Universe.* id-list methods.
- Method parameters are unchanged (an int argument widens on its own);
  only assigning a result into an int needs a caller change.

SpecCheck schema-int-width: 496 -> 0. Remaining warnings are all
model-wider-than-spec extras, deferred.

Integration-test fixtures cast the few small well-known ids back to int.
MIGRATION.md and CHANGELOG.md updated.

+++ ESI.NET/Logic/AllianceLogic.cs
@@ -18,8 +18,8 @@ namespace ESI.NET.Logic
         /// /alliances/
         /// </summary>
         /// <returns></returns>
-        public async Task<EsiResponse<int[]>> All(EsiCallOptions options = null)
-            => await Execute<int[]>(_client, _config, RequestSecurity.Public, HttpMethod.Get, "/alliances/",
+        public async Task<EsiResponse<long[]>> All(EsiCallOptions options = null)
+            => await Execute<long[]>(_client, _config, RequestSecurity.Public, HttpMethod.Get, "/alliances/",
                 options: options);
 
 
@@ -42,8 +42,8 @@ namespace ESI.NET.Logic
         /// </summary>
         /// <param name="alliance_id"></param>
         /// <returns></returns>
-        public async Task<EsiResponse<int[]>> Corporations(int alliance_id, EsiCallOptions options = null)
-            => await Execute<int[]>(_client, _config, RequestSecurity.Public, HttpMethod.Get, "/alliances/{alliance_id}/corporations/",
+        public async Task<EsiResponse<long[]>> Corporations(int alliance_id, EsiCallOptions options = null)
+            => await Execute<long[]>(_client, _config, RequestSecurity.Public, HttpMethod.Get, "/alliances/{alliance_id}/corporations/",
                 replacements: new Dictionary<string, string>()
                 {
                     { "alliance_id", alliance_id.ToString() }
