+++ ESI.NET/Logic/WarsLogic.cs
@@ -17,14 +17,14 @@ namespace ESI.NET.Logic
         /// </summary>
         /// <param name="max_war_id">Only return wars with ID smaller than this</param>
         /// <returns></returns>
-        public async Task<EsiResponse<int[]>> All(long max_war_id = 0, EsiCallOptions options = null)
+        public async Task<EsiResponse<long[]>> All(long max_war_id = 0, EsiCallOptions options = null)
         {
             var parameters = new List<string>();
 
             if (max_war_id > 0)
                 parameters.Add($"max_war_id={max_war_id}");
 
-            var response = await Execute<int[]>(_client, _config, RequestSecurity.Public, HttpMethod.Get, "/wars/",
+            var response = await Execute<long[]>(_client, _config, RequestSecurity.Public, HttpMethod.Get, "/wars/",
                 parameters: parameters.ToArray(),
                 options: options);
