+++ ESI.NET/Models/Character/CorporationHistory.cs
@@ -6,13 +6,13 @@ namespace ESI.NET.Models.Character
     public class CorporationHistory
     {
         [JsonProperty("corporation_id")]
-        public int CorporationId { get; set; }
+        public long CorporationId { get; set; }
 
         [JsonProperty("is_deleted")]
         public bool IsDeleted { get; set; }
 
         [JsonProperty("record_id")]
-        public int RecordId { get; set; }
+        public long RecordId { get; set; }
 
         [JsonProperty("start_date")]
         public DateTime StartDate { get; set; }
