+++ ESI.NET/Models/Contracts/Bid.cs
@@ -6,10 +6,10 @@ namespace ESI.NET.Models.Contracts
     public class Bid
     {
         [JsonProperty("bid_id")]
-        public int BidId { get; set; }
+        public long BidId { get; set; }
 
         [JsonProperty("bidder_id")]
-        public int BidderId { get; set; }
+        public long BidderId { get; set; }
 
         [JsonProperty("date_bid")]
         public DateTime DateBid { get; set; }
