+++ ESI.NET/Models/FactionWarfare/War.cs
@@ -5,9 +5,9 @@ namespace ESI.NET.Models.FactionWarfare
     public class War
     {
         [JsonProperty("faction_id")]
-        public int FactionId { get; set; }
+        public long FactionId { get; set; }
 
         [JsonProperty("against_id")]
-        public int AgainstId { get; set; }
+        public long AgainstId { get; set; }
     }
 }
